[03032] Consolidate UseDataTableUpdates and fix DataTable bugs
- Move UseObservableStream into UseDataTableUpdates in Widgets/DataTables
- Add UseDataTableUpdates wrapper to ViewBase.Hooks for direct access
- Wire UpdateStream through DataTableBuilder and DataTableView
- Fix columnHelpers.ts crash when column type is undefined
- Remove Tendril Plans app (moved to separate module)
